Added a pause to the animation. Press p to pause
It's still a bit weird sometimes.  I made manual grow and shrink tree independent
of the pause, so you can still grow and shrink the tree at any time manually.

However, there seemed to be some weirdness with growing and shrinking while
not paused. Not sure what that's about, but I don't think it should be too bad.