Started to integrate the partial matches.
Still isn't doing well.  If you'll notice in the file:
discoverOronymsForNice.out

It's making a heck of a lot of loops that it doesn't need to, and
not making ones that it does. It's having a hard time getting
past that first phoneme.

I need to do a full trace of the code. By hand. BLeeerrrfff.
TODO:
(1)pay particular attention to lines 202-221 in interpretPhrase()
   in wordBreakdown.cpp.

Time for dinner!