Time for sleep. Doesn't work.
It's having some issues around line 198 or later.

It might be a better idea to roll back to the first version
of the function, and fix that one.

That'd be around commit 3449a7c349f7ef6cfd22911a5ca5a0fe29641750 or so.

I need to keep the stuff from these commits, though:
1fb2ee7b30aa54d710a6f54e8b04649ca1b274a7
861c79e1fe6e279ad9b82accd724038ff6894fbd

But now, sleep. I shall dream disappointed dreams.