Added phone-tree representation
I determined that the best way that my GUI could interact with
the back end code was if the back end code provided a phone "tree"
of sorts.
It's not actually a tree. It's just a vector of sets, where each
set contains the possible phonemes for that place in the
phonetic sequence.

EX:
 Given: a nice
 Returns:   vector[0]: { e, @, A }
            vector[1]: { n }
            vector[2]: { aI, i }
            vector[3]: { s }

This will make branching easier

NEXT:
--create GUI/backend mashup code. YOU CAN DO IT!

THEN:
--incorporate Word frequency into GUI branch radius

Things I want to do:
--INVERSE some mothers (add duplicate phones in for all consonants)