Added freqs for recorded and transcribed phrases
For some reason, excel is very unhappy with one of the recorded
phrases. ONLY ONE, though. It's saying that my call to LOOKUP
is =#N/A or somesuch nonsense.

I'll look into it later.

I think that it's okay that I don't do any special accounting for
nonsense words in my frequency calculation. After all, they're not
actually words.