DEAR LORD IN HEAVEN Protovis is amazing!
The most important files in this commit:

These show a *sunburst diagram* of the frequency of "iced ink" oronyms! :D

I'm using a MULTIPLICATIVE frequency metric, since that's more correct anyway.

TODO:

(1) Revise all frequency metrics to be products instead of sums.
(2) Make protovis of an ice cold hour oronyms!  May want to write a script to generate that.