Comma placement still wrong, but fixable.
It puts commas at the end of EVERY child node, when it should leave
the last child node comma-less.

It doesn't put commas at the end of {} blocks in series. Bummer.

I fixed it manually, and compiled it for the browser.

NOTE:  the HTML file root must be the same name as the variable in
your .js data file.

Problem right now with MULTIPLICATIVE data:

for shorter phrases "why that's insane" vs "why that sin sane".
   the shorter phrases get COMPLETELY obliterated because they don't have
   the multiplicative effect going on.

   I may want to switch back to summation.