Added a bunch of stuff, made Observed sunburst
But now, for some reason, the predictive sunburst doesn't match.

Also, the colors are off, because I added another category,
unpredicted, to the observed sunburst. I'll have to figure out what's
wrong with the predicted sunburst protovis stuff, and then add
that unpredicted category, so that the colors will match on my observed
and predicted sunbursts

Also, I added a bunch of figures for transcriptions per recording.
I need to go through those.

Also, a bunch of other stuff.

When I'm not working on the command line, I always forget to
commit regularly!