Added dbtype enum, plus removed superfluous param requirement
The connectToPhoneticDictionaryDatabase required a string param,
but in the code, it dealt with having the string empty. Instead
of passing the UNISYN filepath every time, or passing an empty
string and having it default to the UNISYN file path, I made it
take a default param value.

Notes on default param: Apparently, you must ONLY set them in the
   .h file, not the .cpp, and not in both.

I ran the ./testWordBreakdown all, and it segfaulted on the drawRect
function. That, I believe, is entirely unrelated to my changes, &
since I'm not even using that anymore, I'm going to remove it from
the default test suite.