Sunburst generation works!
As long as you remove ' and - from the output afterwards, as well as
the first and last lines (debug test text), then it works :D

In this commit, we added the UNISYN sunbursts, with and without the
unpredicted slivers!

I want to cherrypick the COCA Sunburst screenshots and code over to
master, but that can probably be done later. :D